% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_grade_range.R
\name{calculate_grade_range}
\alias{calculate_grade_range}
\title{Calculate grade range}
\usage{
calculate_grade_range(
  completed_ects,
  current_grade,
  remaining_ects,
  max_passing_grade = 1,
  min_passing_grade = 5
)
}
\arguments{
\item{completed_ects}{the number of ECTS credits already completed}

\item{current_grade}{the current grade average}

\item{remaining_ects}{the number of ECTS credits remaining to complete}

\item{max_passing_grade}{the best possible grade}

\item{min_passing_grade}{the worst possible grade}
}
\value{
a vector with the best and worst possible grade averages
}
\description{
This function calculates the best and worst possible grade averages that can
be achieved based on the current grade average, the number of ECTS credits
already completed, and the number of remaining ECTS credits to complete.
}
\examples{

calculate_grade_range(120, 3.5, 60)
calculate_grade_range(120, 3.5, 60, 1, 5)

}
