% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caption_builder.R
\name{caption_builder}
\alias{caption_builder}
\title{Caption builder}
\usage{
caption_builder(
  min_passing_grade,
  max_passing_grade,
  caption_width = 150,
  show_caption = TRUE
)
}
\arguments{
\item{min_passing_grade}{the worst possible grade}

\item{max_passing_grade}{the best possible grade}

\item{caption_width}{text width for the caption. default 150}

\item{show_caption}{whether or not to display the text caption}
}
\value{
a character string
}
\description{
Builds the text caption for the grade range plot. The caption can be optionally turned off on the ggplot2 object which is useful when the text is not needed or rendered well on the image file. Instead, the Shiny app creates a separate text box for the caption.
}
\examples{

caption_builder(min_passing_grade = 5, max_passing_grade = 1)

}
