% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_freedom_units.R
\name{convert_to_freedom_units}
\alias{convert_to_freedom_units}
\title{Convert grades to freedom units}
\usage{
convert_to_freedom_units(grades, max_passing_grade = 4, min_passing_grade = 1)
}
\arguments{
\item{grades}{a vector of grades on the German scale}

\item{max_passing_grade}{the maximum passing grade score in the US}

\item{min_passing_grade}{the minimum passing grade score in the US}
}
\value{
a vector of grades in "freedom units"
}
\description{
This function takes a vector of grade on the German scale and converts them
to the "freedom units" scale i.e. to the equivalent American GPA scale. This
calculation makes use of the so-called "Bavarian formula."
}
\examples{

convert_to_freedom_units(grades = c(1.3, 2.8))

}
