% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradecast-package.R
\docType{package}
\name{gradecast-package}
\alias{gradecast}
\alias{gradecast-package}
\title{gradecast: Calculate and visualize potential school grades}
\description{
Convert between German and US grades/GPA systems. Visualize potential final GPA based on current grades and remaining European Credit Transfer and Accumulation points (ECTS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://neuroshepherd.github.io/gradecast/}
}

}
\author{
\strong{Maintainer}: Pat Callahan \email{patricktcallahan18@gmail.com} (\href{https://orcid.org/0000-0003-1769-7580}{ORCID})

}
\keyword{internal}
