% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradecast-package.R
\docType{package}
\name{gradecast-package}
\alias{gradecast}
\alias{gradecast-package}
\title{gradecast: Calculate and visualize potential school grades}
\description{
Convert between German and US grades/GPA systems. Visualize potential final GPA based on current grades and remaining European Credit Transfer and Accumulation points (ECTS).
}
\keyword{internal}
