% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grade_range.R
\name{plot_grade_range}
\alias{plot_grade_range}
\title{Plot grade range}
\usage{
plot_grade_range(
  completed_ects,
  current_grade,
  remaining_ects,
  top_grade = 1,
  worst_grade = 5
)
}
\arguments{
\item{completed_ects}{the number of ECTS credits already completed}

\item{current_grade}{the current grade average}

\item{remaining_ects}{the number of ECTS credits remaining to complete}

\item{top_grade}{the best possible grade}

\item{worst_grade}{the worst possible grade}
}
\value{
a ggplot2 object
}
\description{
Plot grade range
}
\examples{

# On the German scale
plot_grade_range(30, 1.5, 90)


# Now the same plot, but on the American scale and converting the grades.
# Note the `.` in the `plot_grade_range` function call. This is a placeholder
# for the result of the previous function call.

convert_to_freedom_units(1.5) \%>\%
  plot_grade_range(30, ., 90, top_grade = 4, worst_grade = 1)

}
